backend/pyproject.toml
[project]
name = "Synapse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "PyYAML",
    "sse-starlette",

    # 数据库相关
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",      # SQLite 异步驱动
    "aiomysql>=0.2.0",        # MySQL 异步驱动
    "pymysql>=1.1.0",         # aiomysql 依赖
    "asyncpg>=0.29.0",        # PostgreSQL 异步驱动

    # 配置管理
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
oracle = ["oracledb>=1.4.0"]
dm8 = ["dmPython>=2.3.0"]
